+++ requirements.txt
@@ -1,0 +1,2 @@
+ib-insync==0.9.53
+numpy==1.16.4
